review_api/requirements.txt
appdirs==1.4.4
asgiref==3.8.1
attrs==25.3.0
black==25.1.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
cryptography==44.0.2
defusedxml==0.7.1
Django==5.2
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
djoser==2.3.1
idna==3.10
isort==6.0.1
mypy_extensions==1.1.0
oauthlib==3.2.2
packaging==25.0
pathspec==0.12.1
platformdirs==4.3.7
pycparser==2.22
PyJWT==2.9.0
python-dotenv==1.1.0
python3-openid==3.2.0
requests==2.32.3
requests-oauthlib==2.0.0
social-auth-app-django==5.4.3
social-auth-core==4.5.6
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.4.0
